:art: move helper functions to a separate file and add pagination
